Move & update import to all components tests
